Add data for gene info merger; refactor oncokb tools, and add gene summary reader class
