+++ src/lib/djerba/util/oncokb/constants.py
@@ -26,7 +26,6 @@ LIKELY_ONCOGENIC = 'Likely Oncogenic'
 INCONCLUSIVE = 'Inconclusive'
 UNKNOWN = 'Unknown'
 
-ONCOGENIC_UC = 'ONCOGENIC'
 FDA_APPROVED_LEVELS = [LEVEL_1, LEVEL_2, LEVEL_R1]
 INVESTIGATIONAL_LEVELS = [LEVEL_3A, LEVEL_3B, LEVEL_4, LEVEL_R2]
 THERAPY_LEVELS = [
@@ -51,3 +50,8 @@ ORDERED_LEVELS = [
     INCONCLUSIVE,
     UNKNOWN
 ]
+
+### miscellaneous ###
+
+ALL_CURATED_GENES = '20201126-allCuratedGenes.tsv'
+ONCOGENIC_UC = 'ONCOGENIC'
